+++ arrayjit/lib/ppx_ndarray.ml
@@ -11,7 +11,7 @@ let expr_expander ~loc ~path:_ payload =
       let bindings =
         List.map bindings ~f:(fun vb ->
             let v =
-              (if is_cd dt then translate else translate_dt)
+              translate
                 ~desc_label:vb.pvb_pat vb.pvb_expr
             in
             {
@@ -24,7 +24,7 @@ let expr_expander ~loc ~path:_ payload =
       in
       { payload with pexp_desc = Pexp_let (recflag, bindings, body) }
   | expr ->
-      let expr = (if is_cd dt then translate else translate_dt) expr in
+      let expr = translate expr in
       [%expr
         let open! NTDSL.O in
         [%e expr]]
@@ -36,10 +36,10 @@ let flatten_str ~loc ~path:_ items =
       Ast_helper.Str.include_
         { pincl_mod = Ast_helper.Mod.structure items; pincl_loc = loc; pincl_attributes = [] }
 
-let translate_str ~dt ({ pstr_desc; _ } as str) =
+let translate_str ({ pstr_desc; _ } as str) =
   match pstr_desc with
   | Pstr_eval (expr, attrs) ->
-      let expr = (if is_cd dt then translate else translate_dt) expr in
+      let expr = translate expr in
       let loc = expr.pexp_loc in
       {
         str with
@@ -54,7 +54,7 @@ let translate_str ~dt ({ pstr_desc; _ } as str) =
       let f vb =
         let loc = vb.pvb_loc in
         let v =
-          (if is_cd dt then translate else translate_dt)
+          translate
             ~desc_label:vb.pvb_pat vb.pvb_expr
         in
         {